+++ src/mirax-io.c
@@ -38,29 +38,10 @@ static const char *SLIDE_WSI = "ScanDataLayer_WholeSlide";
 // preview image
 static const char *SLIDE_PREVIEW = "ScanDataLayer_SlidePreview";
 
-const char *concat_wildcard_string_int32(const char *str, int32_t integer) {
-    char *result_string = (char *)malloc(strlen(str) + number_of_digits(integer) + 1);
-    sprintf(result_string, str, integer);
-    return result_string;
-}
-
-const char *concat_wildcard_string_m_int32(const char *str, int32_t integer1, int32_t integer2) {
-    char *result_string =
-        (char *)malloc(strlen(str) + number_of_digits(integer1) + number_of_digits(integer2) + 1);
-    sprintf(result_string, str, integer1, integer2);
-    return result_string;
-}
-
-const char *concat_wildcard_string_string(const char *wildcard_str, const char *replacement) {
-    char *result_string = (char *)malloc(strlen(wildcard_str) + strlen(replacement) + 1);
-    sprintf(result_string, wildcard_str, replacement);
-    return result_string;
-}
-
 // retrieve the file structure of the mirax file from the
 // Slidedat.ini file
 struct mirax_file *get_mirax_file_structure(struct ini_file *ini, int32_t l_count) {
-    // initialize mirrax file and array of associated layers
+    // initialize mirax file and array of associated layers
     struct mirax_file *mirax_file = (struct mirax_file *)malloc(sizeof(struct mirax_file));
 
     struct mirax_layer **layers =
@@ -142,25 +123,6 @@ struct mirax_level *get_level_by_name(struct mirax_layer **layers, const char *l
         }
     }
     return NULL;
-}
-
-// read a signed integer 32 from file stream
-int32_t *read_int32(file_t *fp) {
-    int32_t *buffer = (int32_t *)malloc(sizeof(int32_t));
-
-    if (file_read(buffer, sizeof(*buffer), 1, fp) != 1) {
-        free(buffer);
-        return NULL;
-    }
-
-    return buffer;
-}
-
-// assert a int32_t has a certain value
-// used to skip pointer position on stream
-bool assert_value(file_t *fp, int32_t value) {
-    int32_t *v_to_check = read_int32(fp);
-    return *v_to_check == value;
 }
 
 // read file number, position and size frrom index dat
@@ -316,7 +278,6 @@ int32_t delete_record_from_index_file(const char *filename, int32_t record, int3
     return 0;
 }
 
-// TODO: CHECK IF THIS WORKS UNDER WINDOWS!!
 // duplicate and rename the mirax file file.mrxs and
 // the associated folder with the image data
 // return new path name of image data folder
@@ -411,36 +372,6 @@ struct mirax_layer *delete_level_by_id(struct mirax_layer *layer, int32_t level_
 
     free(layer);
     return temp;
-}
-
-int32_t delete_level_from_mirax_file(struct mirax_file *mirax_file,
-                                     struct mirax_level *level_to_delete) {
-    // find the level id that we want to delete
-    for (int32_t i = 0; i < mirax_file->count_layers; i++) {
-        int32_t level_id = -1;
-        struct mirax_layer *layer = mirax_file->layers[i];
-        for (int32_t j = 0; j < layer->level_count; j++) {
-            struct mirax_level *level = layer->levels[j];
-            if (level_to_delete->id == level->id) {
-                level_id = j;
-                break;
-            }
-        }
-
-        if (level_id != -1) {
-            // remove the level entry and exchange the mirax layer in the array
-            struct mirax_layer *new_layer =
-                (struct mirax_layer *)malloc(sizeof(struct mirax_layer));
-            new_layer = delete_level_by_id(layer, level_id);
-
-            if (new_layer != NULL) {
-                mirax_file->layers[i] = new_layer;
-                mirax_file->all_records_count--;
-                return 0;
-            }
-        }
-    }
-    return -1;
 }
 
 void rename_mirax_file_levels(struct mirax_file *mirax_file, int32_t layer,