+++ src/utils.c
@@ -186,7 +186,7 @@ const char *get_string_between_delimiters(const char *buffer, const char *delimi
 
 void remove_leading_spaces(char *str) {
     char *p = str;
-    int l = strlen(p);
+    int32_t l = strlen(p);
 
     while (isspace(p[l - 1]))
         p[--l] = 0;
@@ -235,7 +235,7 @@ const char *get_filename_from_path(const char *path) {
 }
 
 const char *int32_to_str(int32_t integer) {
-    int length = snprintf(NULL, 0, "%d", integer);
+    int32_t length = snprintf(NULL, 0, "%d", integer);
     char *str = (char *)malloc(length + 1);
     snprintf(str, length + 1, "%d", integer);
     return str;
@@ -273,7 +273,7 @@ const char *add_equals_sign(const char *str1, const char *str2) {
 
 // function to check if str1 contains str2
 bool contains(const char *str1, const char *str2) {
-    int i = 0, j = 0;
+    int32_t i = 0, j = 0;
 
     while (str1[i] != '\0') {
         if (str1[i] == str2[j]) {
@@ -292,8 +292,8 @@ bool contains(const char *str1, const char *str2) {
 }
 
 // function to check how many times str1 contains str2
-int count_contains(const char *str1, const char *str2) {
-    int i = 0, j = 0, count = 0;
+int32_t count_contains(const char *str1, const char *str2) {
+    int32_t i = 0, j = 0, count = 0;
 
     while (str1[i] != '\0') {
         if (str1[i] == str2[j]) {
@@ -390,7 +390,7 @@ int32_t copy_directory(const char *src, const char *dest) {
 // determine wether the operating system
 // is big or little endian
 bool is_system_big_endian() {
-    int n = 1;
+    int32_t n = 1;
     if (*(char *)&n == 1) {
         return true;
     }
@@ -430,4 +430,84 @@ int32_t bytes_to_int(unsigned char *buffer, int32_t size) {
         shift += 8;
     }
     return ret;
+}
+
+// find size up to substring in file
+int32_t get_size_to_substring(file_t *fp, char *substring) {
+
+    file_seek(fp, 0, SEEK_END);
+    long file_length = file_tell(fp);
+    char *buffer = (char *)malloc(file_length);
+    file_seek(fp, 0, SEEK_SET);
+
+    if (file_read(buffer, file_length, 1, fp) != 1) {
+        free(buffer);
+        fprintf(stderr, "Error: Could not read file.\n");
+        return -1;
+    }
+
+    // finds size
+    char *ret = strstr(buffer, substring);
+    int32_t size = ret - buffer;
+
+    file_seek(fp, 0, SEEK_SET);
+    free(buffer);
+
+    return size;
+}
+
+// check if file contains specific value
+int32_t file_contains_value(file_t *fp, char *value) {
+
+    file_seek(fp, 0, SEEK_END);
+    long size = file_tell(fp);
+    char *buffer = (char *)malloc(size);
+    file_seek(fp, 0, SEEK_SET);
+
+    if (file_read(buffer, size, 1, fp) != 1) {
+        free(buffer);
+        fprintf(stderr, "Error: Could not read file.\n");
+        return -1;
+    }
+
+    // searches for value
+    if (contains(buffer, value)) {
+        free(buffer);
+        return 1;
+    }
+
+    free(buffer);
+    return -1;
+}
+
+const char *concat_wildcard_string_int32(const char *str, int32_t integer) {
+    char *result_string = (char *)malloc(strlen(str) + number_of_digits(integer) + 1);
+    sprintf(result_string, str, integer);
+    return result_string;
+}
+
+const char *concat_wildcard_string_m_int32(const char *str, int32_t integer1, int32_t integer2) {
+    char *result_string =
+        (char *)malloc(strlen(str) + number_of_digits(integer1) + number_of_digits(integer2) + 1);
+    sprintf(result_string, str, integer1, integer2);
+    return result_string;
+}
+
+// read a signed integer 32 from file stream
+int32_t *read_int32(file_t *fp) {
+    int32_t *buffer = (int32_t *)malloc(sizeof(int32_t));
+
+    if (file_read(buffer, sizeof(*buffer), 1, fp) != 1) {
+        free(buffer);
+        return NULL;
+    }
+
+    return buffer;
+}
+
+// assert a int32_t has a certain value
+// used to skip pointer position on stream
+bool assert_value(file_t *fp, int32_t value) {
+    int32_t *v_to_check = read_int32(fp);
+    return *v_to_check == value;
 }