+++ src/utils.h
@@ -4,7 +4,7 @@
 #include "defines.h"
 #include "file-api.h"
 
-// string operations
+// string and int operations
 char **str_split(char *a_str, const char a_delim);
 
 const char *get_filename_ext(const char *filename);
@@ -42,6 +42,18 @@ const char *add_equals_sign(const char *str1, const char *str2);
 
 bool contains(const char *str1, const char *str2);
 
+int32_t get_size_to_substring(file_t *fp, char *substring);
+
+int32_t file_contains_value(file_t *fp, char *value);
+
+const char *concat_wildcard_string_int32(const char *str, int32_t integer);
+
+const char *concat_wildcard_string_m_int32(const char *str, int32_t integer1, int32_t integer2);
+
+int32_t *read_int32(file_t *fp);
+
+bool assert_value(file_t *fp, int32_t value);
+
 // file operations
 const char *duplicate_file(const char *filename, const char *new_file_name,
                            const char *file_extension);
@@ -63,7 +75,7 @@ uint64_t _swap_uint64(uint64_t value);
 
 char *skip_first_and_last_char(char *value);
 
-int count_contains(const char *str1, const char *str2);
+int32_t count_contains(const char *str1, const char *str2);
 
 int32_t bytes_to_int(unsigned char *buffer, int32_t size);